Export Command
New command: 'export'. It does what you would expect, exports a query to
a file. The export command take input which is processed like a search
function, then prints the results set, and results statistics to a file
('d/m/y DB Export.txt').

- 'insert' is no longer flag order specific. It never should have been
in the first place, but it was easier to implement...
- Sorting ('-s' or '--sort' flag) now can take any number of columns.
The results will be sorted in order of column arguments after the flag.
Again, should have been there to begin with, but it's there now.
- README updated to reflect new changes
- dataprint.py recieved some interesting changes to better accommodate
'export'. G'bless those lambda functions, what a great feature of
python.

No bug fixes this update, but probably introduced a lot of new bugs. Oh
well. New code is a little rough, will be probably be revised in the
future...

Version change: Terminus v1.6