TSV Command and Plot Changes
New command 'tsv' exports a query as a .tsv file. The tsv format was
choosen over csv, because many book titles have commas in their titles.
I may change upload to work with tsv files in future as a result of this
change.

Additional changes:
- plot command no longer needs a '-w' flag to specify a wher clause. Any
flag which is not a -X or -Y flag is interpretted as a normal query
flag.
- Fixed a spacing bug in Y axis aggregated plot commands, and removed a
function which was never called.
- Adjusted text spacing on the title screen
- Updated help text, and README to reflect new command, and plot changes

Version change: Terminus v2.1