Photo album
Added an album of a few of the more visual elements of Terminus. The
photos are from v2.4, but it looks the same as in this version (and
probably future versions).

Also, fixed a bug involving a weird edge case in the quote printing
where lines wouldn't be broken properly. A pointless for loop was cut
too.

Version change: Terminus v2.5